requirements.txt
async-generator==1.10
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
dnspython==2.3.0
docopt==0.6.2
exceptiongroup==1.1.1
Flask==2.2.3
Flask-Cors==3.0.10
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.3
outcome==1.2.0
packaging==23.1
pandas==2.0.1
pipreqs==0.4.11
pymongo==4.3.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
selenium==4.9.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
tzdata==2023.3
urllib3==1.26.14
webdriver-manager==3.8.6
Werkzeug==2.2.3
wsproto==1.2.0
yarg==0.1.9
zipp==3.15.0
